#1539: Got rid of circular import by moving function to calling code file.
